# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stunting",
        "description": "Dokumentasi API untuk Sistem Stunting",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/kalkulator-fuzzy": {
            "get": {
                "tags": [
                    "Kalkulator Fuzzy"
                ],
                "summary": "Ambil riwayat perhitungan fuzzy pengguna",
                "operationId": "010cde55e7bad4c4194270fc09500c78",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil riwayat perhitungan fuzzy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ambil data"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama_bayi": {
                                                        "type": "string",
                                                        "example": "Bayi 1"
                                                    },
                                                    "kondisi_anak": {
                                                        "type": "string",
                                                        "example": "Normal"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-09-01T00:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil riwayat perhitungan fuzzy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Gagal mendapatkankan data"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Kalkulator Fuzzy"
                ],
                "summary": "Hitung data kalkulator fuzzy",
                "operationId": "e3c69d9e92bf21455f19871ba9519a91",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_bayi",
                                    "jenis_kelamin",
                                    "usia",
                                    "berat_badan",
                                    "tinggi_badan"
                                ],
                                "properties": {
                                    "nama_bayi": {
                                        "type": "string",
                                        "example": "Bayi 1"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string",
                                        "enum": [
                                            "Laki - Laki",
                                            "Perempuan"
                                        ],
                                        "example": "Laki - Laki"
                                    },
                                    "usia": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "berat_badan": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.5
                                    },
                                    "tinggi_badan": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 75.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data fuzzy berhasil dihitung",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data fuzzy berhasil dihitung."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "z_score_bbu": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": -1.2
                                                },
                                                "z_score_tbu": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 0.8
                                                },
                                                "z_score_bbtb": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1.1
                                                },
                                                "fuzzy_bbu": {
                                                    "type": "string",
                                                    "example": "Normal"
                                                },
                                                "fuzzy_tbu": {
                                                    "type": "string",
                                                    "example": "Pendek"
                                                },
                                                "fuzzy_bbtb": {
                                                    "type": "string",
                                                    "example": "Normal"
                                                },
                                                "kondisi_anak": {
                                                    "type": "string",
                                                    "example": "Normal"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Kesalahan validasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Kesalahan validasi"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gagal menghitung data fuzzy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Gagal menghitung data fuzzy"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kalkulator-fuzzy/{id}": {
            "get": {
                "tags": [
                    "Kalkulator Fuzzy"
                ],
                "summary": "Ambil data perhitungan fuzzy berdasarkan ID",
                "operationId": "432c54dbf887f7bb6613241413c5e559",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Kalkulator Fuzzy",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ambil data"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nama_bayi": {
                                                    "type": "string",
                                                    "example": "Bayi 1"
                                                },
                                                "kondisi_anak": {
                                                    "type": "string",
                                                    "example": "Normal"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-01T00:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Gagal mendapatkankan data"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gagal mendapatkan data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Gagal mendapatkankan data"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Laravel JWT authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}